Reorganization of transaction related modules
