+++ wallet/src/Ergvein/Wallet/Transaction.hs
@@ -1,0 +1,9 @@
+module Ergvein.Wallet.Transaction(
+    module Ergvein.Wallet.Transaction.Get
+  , module Ergvein.Wallet.Transaction.Util
+  , module Ergvein.Wallet.Transaction.View
+  ) where
+
+import Ergvein.Wallet.Transaction.Get
+import Ergvein.Wallet.Transaction.Util
+import Ergvein.Wallet.Transaction.View
